Move Pitches to the core, removes Note enumeration
