+++ jarduino.core/src/main/java/org/sintef/jarduino/Pitches.java
@@ -1,4 +1,4 @@
-package org.sintef.jarduino.examples.digital;
+package org.sintef.jarduino;
 
 
 public interface Pitches {
