+++ jarduino.samples/src/main/java/org/sintef/jarduino/examples/digital/ToneMultiple.java
@@ -19,6 +19,7 @@ package org.sintef.jarduino.examples.digital;
 
 import org.sintef.jarduino.DigitalPin;
 import org.sintef.jarduino.JArduino;
+import org.sintef.jarduino.Pitches;
 import org.sintef.jarduino.comm.Serial4JArduino;
 
 public class ToneMultiple extends JArduino implements Pitches {
